tests: split py36 tests
