+++ tox.ini
@@ -18,7 +18,8 @@ deps = -e.[test]
 ; Prevent installing issues: https://github.com/ContinuumIO/anaconda-issues/issues/542
 commands =
     pip install -U setuptools
-    pytest --cov-report=term-missing --cov=gql tests {posargs}
+    py{27,35}: pytest --cov-report=term-missing --cov=gql tests {posargs}
+    py{36,37,38}: pytest --cov-report=term-missing --cov=gql tests tests_py36 {posargs}
 
 [testenv:flake8]
 basepython = python3.8
